Introduce `subspace_node::commands::run` module for running node in normal mode (code is moved, no changes other than imports)
