+++ crates/subspace-node/src/commands.rs
@@ -1,0 +1,3 @@
+mod run;
+
+pub use run::run;
